backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
psycopg2-binary==2.9.9
python-multipart==0.0.6
requests==2.31.0
nltk==3.8.1
scikit-learn==1.3.2
numpy==1.26.4
librosa==0.10.0
transformers==4.36.2
torch==2.2.0
python-multipart==0.0.6
